REFACTOR: 설정, 스플래시, 약관, 웹뷰, 탈퇴 화면 경로 수정
