+++ presentation/src/main/java/com/threegap/bitnagil/presentation/screen/setting/contract/SettingSideEffect.kt
@@ -1,0 +1,6 @@
+package com.threegap.bitnagil.presentation.screen.setting.contract
+
+sealed interface SettingSideEffect {
+    data object NavigateToLogin : SettingSideEffect
+    data object NavigateToWithdrawal : SettingSideEffect
+}
