+++ presentation/src/main/java/com/threegap/bitnagil/presentation/screen/withdrawal/contract/WithdrawalSideEffect.kt
@@ -1,0 +1,6 @@
+package com.threegap.bitnagil.presentation.screen.withdrawal.contract
+
+sealed interface WithdrawalSideEffect {
+    data object NavigateToBack : WithdrawalSideEffect
+    data object NavigateToLogin : WithdrawalSideEffect
+}
