--- presentation/src/main/java/com/threegap/bitnagil/presentation/setting/contract/SettingState.kt
@@ -1,23 +1,0 @@
-package com.threegap.bitnagil.presentation.setting.contract
-
-import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
-data class SettingState(
-    val useServiceAlarm: Boolean,
-    val usePushAlarm: Boolean,
-    val version: String,
-    val loading: Boolean,
-    val logoutConfirmDialogVisible: Boolean,
-) : Parcelable {
-    companion object {
-        val Init = SettingState(
-            useServiceAlarm = false,
-            usePushAlarm = false,
-            version = "",
-            loading = false,
-            logoutConfirmDialogVisible = false,
-        )
-    }
-}
