--- presentation/src/main/java/com/threegap/bitnagil/presentation/terms/contract/TermsAgreementSideEffect.kt
@@ -1,8 +1,0 @@
-package com.threegap.bitnagil.presentation.terms.contract
-
-sealed interface TermsAgreementSideEffect {
-    data object NavigateToTermsOfService : TermsAgreementSideEffect
-    data object NavigateToPrivacyPolicy : TermsAgreementSideEffect
-    data object NavigateToOnBoarding : TermsAgreementSideEffect
-    data object NavigateToBack : TermsAgreementSideEffect
-}
