--- presentation/src/main/java/com/threegap/bitnagil/presentation/terms/contract/TermsAgreementState.kt
@@ -1,23 +1,0 @@
-package com.threegap.bitnagil.presentation.terms.contract
-
-data class TermsAgreementState(
-    val isLoading: Boolean,
-    val agreedTermsOfService: Boolean,
-    val agreedPrivacyPolicy: Boolean,
-    val agreedOverFourteen: Boolean,
-) {
-    val isAllAgreed: Boolean
-        get() = agreedTermsOfService && agreedPrivacyPolicy && agreedOverFourteen
-
-    val submitEnabled: Boolean
-        get() = !isLoading && isAllAgreed
-
-    companion object {
-        val INIT = TermsAgreementState(
-            isLoading = false,
-            agreedTermsOfService = false,
-            agreedPrivacyPolicy = false,
-            agreedOverFourteen = false,
-        )
-    }
-}
