--- presentation/src/main/java/com/threegap/bitnagil/presentation/webview/BitnagilWebViewScreen.kt
@@ -1,96 +1,0 @@
-package com.threegap.bitnagil.presentation.webview
-
-import android.view.ViewGroup
-import android.webkit.WebResourceRequest
-import android.webkit.WebSettings
-import android.webkit.WebView
-import android.webkit.WebViewClient
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.statusBarsPadding
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.DisposableEffect
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.viewinterop.AndroidView
-import com.threegap.bitnagil.designsystem.BitnagilTheme
-import com.threegap.bitnagil.designsystem.component.block.BitnagilTopBar
-
-@Composable
-fun BitnagilWebViewScreen(
-    title: String,
-    url: String,
-    onBackClick: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    val context = LocalContext.current
-    val webView = remember {
-        WebView(context).apply {
-            webViewClient = object : WebViewClient() {
-                override fun shouldOverrideUrlLoading(
-                    view: WebView?,
-                    request: WebResourceRequest?,
-                ): Boolean {
-                    val requestUrl = request?.url?.toString()
-                    return requestUrl?.contains("notion.site") != true
-                }
-            }
-
-            settings.apply {
-                javaScriptEnabled = true
-                domStorageEnabled = true
-                allowFileAccess = false
-                allowContentAccess = false
-                mixedContentMode = WebSettings.MIXED_CONTENT_NEVER_ALLOW
-                setSupportZoom(false)
-            }
-
-            layoutParams = ViewGroup.LayoutParams(
-                ViewGroup.LayoutParams.MATCH_PARENT,
-                ViewGroup.LayoutParams.MATCH_PARENT,
-            )
-        }
-    }
-
-    DisposableEffect(webView) {
-        onDispose {
-            webView.destroy()
-        }
-    }
-
-    Column(
-        modifier = modifier
-            .fillMaxSize()
-            .background(BitnagilTheme.colors.white)
-            .statusBarsPadding(),
-    ) {
-        BitnagilTopBar(
-            title = title,
-            showBackButton = true,
-            onBackClick = onBackClick,
-        )
-
-        AndroidView(
-            modifier = Modifier.fillMaxSize(),
-            factory = { webView },
-            update = { view ->
-                if (view.url != url && url.isNotBlank()) {
-                    view.loadUrl(url)
-                }
-            },
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun BitnagilWebViewScreenPreview() {
-    BitnagilWebViewScreen(
-        title = "약관 동의",
-        url = "",
-        onBackClick = {},
-    )
-}
