--- presentation/src/main/java/com/threegap/bitnagil/presentation/withdrawal/contract/WithdrawalSideEffect.kt
@@ -1,6 +1,0 @@
-package com.threegap.bitnagil.presentation.withdrawal.contract
-
-sealed interface WithdrawalSideEffect {
-    data object NavigateToBack : WithdrawalSideEffect
-    data object NavigateToLogin : WithdrawalSideEffect
-}
